# api/openapi.yaml
openapi: 3.0.3
info:
  title: messageFeed API
  version: 0.2.0
  description: Phase two and early phase four API contract for the feed reading workflow.
servers:
  - url: http://localhost:60001
    description: Local API server
paths:
  /api/v1/sources:
    get:
      summary: List subscription sources
      responses:
        '200':
          description: Subscription sources
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Source'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      summary: Create subscription source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '201':
          description: Created source
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Source'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sources/{id}:
    patch:
      summary: Update subscription source
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        '200':
          description: Updated source
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Source'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sources/{id}/fetch:
    post:
      summary: Fetch one source immediately
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Fetch result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FetchSourceResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/source-fetches:
    post:
      summary: Fetch all active sources
      responses:
        '200':
          description: Batch fetch result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FetchSourcesResult'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/source-catalogs:
    get:
      summary: List source catalog entries
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Source catalog entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SourceCatalogList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/source-catalogs/search:
    get:
      summary: Search source catalog entries
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Source catalog entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SourceCatalogList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sources/import/catalog:
    post:
      summary: Import sources from catalog entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - catalog_ids
              properties:
                catalog_ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                    minimum: 1
      responses:
        '200':
          $ref: '#/components/responses/ImportSourceResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sources/import/urls:
    post:
      summary: Import sources from feed URLs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - urls
              properties:
                urls:
                  type: array
                  items:
                    type: string
                    format: uri
      responses:
        '200':
          $ref: '#/components/responses/ImportSourceResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sources/import/opml:
    post:
      summary: Import sources from OPML file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          $ref: '#/components/responses/ImportSourceResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sources/import-jobs:
    get:
      summary: List source import jobs
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Source import jobs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SourceImportJobList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
    ServiceUnavailable:
      description: Service dependency is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
...
    ImportSourceResult:
      description: Import source result
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/APIResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ImportSourceResult'
  schemas:
    APIResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 0 for success; HTTP-like code for current errors.
          example: 0
        message:
          type: string
          example: success
        request_id:
          type: string
        trace_id:
          type: string
          description: OpenTelemetry trace id for correlating API responses with logs and traces.
        data:
          nullable: true
    CreateSourceRequest:
      type: object
      required:
        - url
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        url:
          type: string
          format: uri
        fetch_interval_seconds:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
        weight:
          type: integer
    UpdateSourceRequest:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        url:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/SourceStatus'
        fetch_interval_seconds:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
        weight:
          type: integer
    Source:
      type: object
      required:
        - id
        - user_id
        - name
        - type
        - url
        - normalized_url
        - status
        - fetch_interval_seconds
        - tags
        - weight
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        url:
          type: string
          format: uri
        normalized_url:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/SourceStatus'
        fetch_interval_seconds:
          type: integer
        tags:
          type: array
          items:
            type: string
        weight:
          type: integer
        last_fetched_at:
          type: string
          format: date-time
          nullable: true
        last_fetch_status:
          type: string
          enum:
            - success
            - failed
        last_fetch_error:
          type: string
        last_fetch_duration_ms:
          type: integer
          nullable: true
        last_fetch_item_count:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SourceType:
      type: string
      enum:
        - rss
        - atom
        - json_feed
    SourceStatus:
      type: string
      enum:
        - active
        - inactive
    FetchSourceResult:
      type: object
      required:
        - source
        - item_count
        - created_count
        - updated_count
      properties:
        source:
          $ref: '#/components/schemas/Source'
        item_count:
          type: integer
        created_count:
          type: integer
        updated_count:
          type: integer
    FetchSourcesResult:
      type: object
      required:
        - requested_count
        - success_count
        - failure_count
        - sources
        - errors
      properties:
        requested_count:
          type: integer
        success_count:
          type: integer
        failure_count:
          type: integer
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FetchSourceError'
...
    SourceCatalogList:
      type: object
      required:
        - entries
        - total
        - limit
        - offset
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/SourceCatalogEntry'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...
    SourceImportJobList:
      type: object
      required:
        - jobs
        - total
        - limit
        - offset
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/SourceImportJob'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...